Fix newline at end of file warnings in model sources
Added missing newlines at the end of multiple source and header files in the model directory to resolve compiler warnings. Also updated dependency files for foc_encoder and foc_pwm to reflect changes in included headers. Removed and updated some object and output files as part of the build process.